Change lineSeparator in files, and int to BIGINT

+++ src/main/resources/init_db.sql
@@ -1,7 +1,7 @@
 CREATE DATABASE `taxi_service_db` /*!40100 DEFAULT CHARACTER SET utf8mb3 */ /*!80016 DEFAULT ENCRYPTION='N' */;
 
 CREATE TABLE `manufacturers` (
-  `id` int NOT NULL AUTO_INCREMENT,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
   `name` varchar(255) DEFAULT NULL,
   `counrty` varchar(20) DEFAULT NULL,
   PRIMARY KEY (`id`)
